% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{calculateManhattanDistance}
\alias{calculateManhattanDistance}
\title{Calculates the manhattan distance.}
\usage{
calculateManhattanDistance(data)
}
\arguments{
\item{data}{2D structure with n rows and 2 colums that represents the coordinate in a plane.}
}
\value{
Matrix, of size \eqn{nrow(data) \times nrow(data)}, with the distance between each the pair of points.
}
\description{
\code{calculateManhattanDistance} Calculates the manhattan distance between each pair of nodes.
}
\examples{
fileInput <- system.file("testdata", "chessboard.csv", package="Irescale")
data<-loadChessBoard(fileInput)
distM<-calculateManhattanDistance(data$data)
}
