% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{coor}
\alias{coor}
\title{Transforms a x,y position in a cartesian plane into a position in a 1D array.}
\usage{
coor(i, j, size)
}
\arguments{
\item{i, }{the value of the row.}

\item{j, }{the value of the column.}

\item{size, }{the maximum between row and columns of the matrix.}
}
\value{
an integer value that represents the position in the array.
}
\description{
\code{coor} Transforms a x,y position in a cartesian plane into a position in a 1D array.
}
\examples{
pos<-coor(1,1,10)
}
