% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{expectedValueI}
\alias{expectedValueI}
\title{Calculates the expected value for local I}
\usage{
expectedValueI(W)
}
\arguments{
\item{W}{Weighted Distance Matrix.}
}
\value{
Expected Value
}
\description{
\code{expectedValueI} Calculates the expected value for local I
}
\examples{
W<-matrix(1:100,nrow=10,ncol=10)
evI<-expectedValueI(W)
}
