% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{loadSatelliteImage}
\alias{loadSatelliteImage}
\title{Loads a Satellite image in PNG format}
\usage{
loadSatelliteImage(fileName)
}
\arguments{
\item{fileName}{file's name and path to the file}
}
\value{
An cimg object in gray scale.
}
\description{
\code{loadSatelliteImage} Loads a Satellite image in PNG format. It does not matter the number of chanells it will return it in grayscale (One channel)
}
\examples{
fileInput <- system.file("testdata", "imageGray.png", package="Irescale")
img<-loadSatelliteImage(fileInput)
}
