% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{nullDristribution}
\alias{nullDristribution}
\title{Calculate a distribution of how the var of interest is correlated to a}
\usage{
nullDristribution(data, numReplicates)
}
\arguments{
\item{data}{the distance matrix. Altough the equation asks for weighted distant matrix, the paramenter that is required is only the distance matrix because this procedure calculate calculates the weighted distance mantrix by itself.}

\item{numReplicates}{the variable of interest to calculate Moran's I.}
}
\value{
Histogram and the vector of correlations between residuals
}
\description{
\code{nullDristribution} Calculate a linear regression between variable of interest and latitude, longitude and latitude*longitude. The residuals of this data set is calculated
The variable of interest is shuffle by numReplicates times and each time the linear regression and residuals are calculated.
At each interation the correlation between the original residuals and the shuffle residuals is calculated
This vector os correlations is returned and plot it as histogram.
}
\examples{
inputFileName<-system.file("testdata", "chen.csv", package="Irescale")
input<-loadFile(inputFileName)
c<-nullDristribution(input,1000)
}
