% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{procrustes}
\alias{procrustes}
\title{Procrustes distance between two surfaces}
\usage{
procrustes(U, V)
}
\arguments{
\item{U}{Vector of the first surface.}

\item{V}{Vector of the second surface.}
}
\value{
Procrustes distance
}
\description{
\code{procrustes} Procrustes distance between two surfaces. The Procrustes distance is used to quantify the similarity or dissimilarity of (3-dimensional) shapes, and extensively used in biological morphometrics.
}
