% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{standardizedByColumn}
\alias{standardizedByColumn}
\title{Scales a matrix by column.}
\usage{
standardizedByColumn(M)
}
\arguments{
\item{M}{Matrix to be scaled by column.}
}
\value{
a matrix scaled by column.
}
\description{
\code{standardizedByColumn} It considers each column independently to scale them.
}
