% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectifiedI.R
\name{transformImageToMatrix}
\alias{transformImageToMatrix}
\title{Transforms the image to a matrix.}
\usage{
transformImageToMatrix(im)
}
\arguments{
\item{im}{cimg object.}
}
\description{
\code{transformImageToMatrix} transforms the image into a 2D matrix.
}
\examples{
fileInput <- system.file("testdata", "imageGray.png", package="Irescale")
img<-loadSatelliteImage(fileInput)
data<-transformImageToList(img)
}
